sensors/encrypted_traffic_analysis/requirements.txt
scapy[complete]==2.5.0
kafka-python==2.0.2
python-dotenv==1.0.1
cryptography>=42.0